JK: saved work

+++ MK6HubService.cpp
@@ -186,7 +186,7 @@ void MK6HubService::setChannel(uint8_t channel, int8_t value) {
 
     MICROBIT_DEBUG_DMESG("MK6HubService::setChannel");
 
-    channelValues[channel] = (uint8_t)(value + 0x80);
+    channelValues[channel] = (uint8_t)((int16_t)value + 0x80);
 }
 
 
@@ -236,7 +236,7 @@ void MK6HubService::connect() {
 
 void MK6HubService::setChannel(uint8_t channel, int8_t value) {
 
-    channelValues[channel] = (uint8_t)(value + 0x80);
+    channelValues[channel] = (uint8_t)((int16_t)value + 0x80);
 }
